Clean up style and comments in test
